Moved the backend function files to a new folder: 'api'
