+++ ecosphere-goods/src/components/pages/dashboard/product-catalog/ProductCatalogContext.js
@@ -1,11 +1,11 @@
 import React, { createContext, useEffect, useState } from 'react';
-import { addNewProduct, fetchProductById, fetchProducts, updateProduct } from '../../../../utility-functions/productHandling';
+import { addNewProduct, fetchProductById, fetchProducts, updateProduct } from '../../../../api/productHandling';
+import { deleteProduct } from '../../../../api/productHandling';
 import { useSelector } from 'react-redux';
 import { selectUser } from '../../../../features/userSlice';
 import { toast } from "react-toastify";
 import { ref, uploadString, getDownloadURL } from "firebase/storage";
 import { storage } from "../../../../firebase";
-import { deleteProduct } from '../../../../utility-functions/productHandling';
 
 const ProductCatalogContext = createContext();
