+++ ecosphere-goods/src/components/pages/user-portal-page/user-credentials/SignInPanel.js
@@ -4,7 +4,7 @@ import { Box } from '@mui/material';
 import PasswordTextField from './PasswordTextField';
 import BackToBrowsingButton from '../../../utility/general-buttons/BackToBrowsingButton';
 import { useNavigate } from 'react-router-dom';
-import { handleSignIn } from '../../../../utility-functions/userAuth';
+import { handleSignIn } from '../../../../api/userAuth';
 import { toast } from 'react-toastify';
 import UncontainedButton from '../../../utility/general-buttons/UncontainedButton';
